pyproject.toml
[tool.poetry]
name = "ai-agent-service"
version = "0.1.0"
description = "AI Agent Gateway Service with DDD Architecture"
authors = ["dev"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
fastapi = "0.115.0"
uvicorn = {version = "0.32.0", extras = ["standard"]}
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "2.6.0"
litellm = ">=1.52.8"
bcrypt = ">=4.0.0"
PyJWT = ">=2.8.0"
sqlalchemy = "2.0.36"
alembic = "1.14.0"
asyncpg = "0.30.0"
pgvector = "0.3.6"
redis = "5.2.0"
hiredis = "3.0.0"
python-dotenv = "1.0.1"
httpx = "0.28.0"
tenacity = "9.0.0"
python-json-logger = "3.2.1"
prometheus-client = "0.21.0"
greenlet = "^3.3.2"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.4"
pytest-asyncio = ">=0.24.0"
pytest-cov = ">=6.0.0"
black = ">=24.10.0"
ruff = ">=0.8.4"
mypy = ">=1.13.0"
isort = ">=5.13.2"
autoflake = ">=2.3.1"
pre-commit = ">=3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip_gitignore = true

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true

[[tool.mypy.overrides]]
module = ["tests.*", "alembic.*"]
ignore_errors = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
